% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conesa.R
\name{conesa}
\alias{conesa}
\title{Cone Surface Area}
\usage{
conesa(radius = 0, height = 0)
}
\arguments{
\item{radius}{of cone. Defaults to 0.}

\item{height}{of cone. Defaults to 0.}
}
\description{
This function allows the user to determine the surface area of a right circular cone from its radius and height.
}
\keyword{area,}
\keyword{cone}
\keyword{pi,}
\keyword{radius,}
\keyword{surface}
