% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conevol.R
\name{conevol}
\alias{conevol}
\title{Cone Volume}
\usage{
conevol(radius = 0, height = 0)
}
\arguments{
\item{radius}{of cone. Defaults to 0.}

\item{height}{of cone. Defaults to 0.}
}
\description{
This function allows the user to determine the volume of a right circular cone from its radius and height.
}
\keyword{cone}
\keyword{pi,}
\keyword{radius,}
\keyword{volume,}
