% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radarea.R
\name{radarea}
\alias{radarea}
\title{Radius to Area}
\usage{
radarea(radius = 0)
}
\arguments{
\item{radius}{of circle. Defaults to 0.}
}
\description{
This function allows the user to determine the area of a circle from its radius
}
\keyword{area}
\keyword{pi,}
\keyword{radius,}
