% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radcirc.R
\name{radcirc}
\alias{radcirc}
\title{Radius to Circumference}
\usage{
radcirc(radius = 0)
}
\arguments{
\item{radius}{of circle. Defaults to 0.}
}
\description{
This function allows the user to determine the circumference of a circle from the radius
}
\keyword{pi}
\keyword{radius,}
