% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spheresa.R
\name{spheresa}
\alias{spheresa}
\title{Radius to Sphere Surface Area}
\usage{
spheresa(radius = 0)
}
\arguments{
\item{radius}{of sphere. Defaults to 0.}
}
\description{
This function allows the user to determine the surface area of a sphere based on its radius
}
\keyword{area,}
\keyword{pi,}
\keyword{radius,}
\keyword{sphere}
\keyword{surface}
