% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spherevol.R
\name{spherevol}
\alias{spherevol}
\title{Sphere volume}
\usage{
spherevol(radius = 0)
}
\arguments{
\item{radius}{of sphere. Defaults to 0.}
}
\description{
This function allows the user to determine the volume of a sphere from its radius.
}
\keyword{pi,}
\keyword{radius,}
\keyword{sphere}
\keyword{volume,}
